requirements.txt
load_dotenv
ollama
crewai
streamlit